app/requirements.txt
azure-iot-device
python-dotenv
rpi_ws281x
adafruit-circuitpython-neopixel